+++ src/blockstore-utils.js
@@ -1,8 +1,9 @@
 'use strict'
 
-const base32 = require('base32.js')
 const { Key } = require('interface-datastore')
 const CID = require('cids')
+const multibase = require('multibase')
+const errcode = require('err-code')
 
 /**
  * Transform a cid to the appropriate datastore key.
@@ -11,19 +12,22 @@ const CID = require('cids')
  * @returns {Key}
  */
 exports.cidToKey = cid => {
-  const enc = new base32.Encoder()
-  return new Key('/' + enc.write(cid.buffer).finalize(), false)
+  if (!CID.isCID(cid)) {
+    throw errcode(new Error('Not a valid cid'), 'ERR_INVALID_CID')
+  }
+
+  return new Key('/' + multibase.encode('base32', cid.multihash).toString().slice(1).toUpperCase(), false)
 }
 
 /**
  * Transform a datastore Key instance to a CID
+ * As Key is a multihash of the CID, it is reconstructed using IPLD's RAW codec.
+ * Hence it is highly probable that stored CID will differ from a CID retrieved from blockstore.
  *
  * @param {Key} key
  * @returns {CID}
  */
 exports.keyToCid = key => {
   // Block key is of the form /<base32 encoded string>
-  const decoder = new base32.Decoder()
-  const buff = decoder.write(key.toString().slice(1)).finalize()
-  return new CID(Buffer.from(buff))
+  return new CID(1, 'raw', multibase.decode('b' + key.toString().slice(1).toLowerCase()))
 }