+++ test/api-addr-test.js
@@ -1,9 +1,8 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
+const { Buffer } = require('buffer')
+const { expect } = require('./utils/chai')
 const apiAddr = require('../src/api-addr')
 
 module.exports = () => {