+++ test/datastore-test.js
@@ -2,15 +2,14 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
-const _ = require('lodash')
+const { Buffer } = require('buffer')
+const { expect } = require('./utils/chai')
+const range = require('just-range')
 const Key = require('interface-datastore').Key
 
 module.exports = (repo) => {
   describe('datastore', () => {
-    const dataList = _.range(100).map((i) => Buffer.from(`hello-${i}-${Math.random()}`))
+    const dataList = range(100).map((i) => Buffer.from(`hello-${i}-${Math.random()}`))
     const data = Buffer.from('hello world')
     const b = new Key('hello')
 
@@ -25,7 +24,7 @@ module.exports = (repo) => {
 
       it('massive multiwrite', async function () {
         this.timeout(15000) // add time for ci
-        await Promise.all(_.range(100).map((i) => {
+        await Promise.all(range(100).map((i) => {
           return repo.datastore.put(new Key('hello' + i), dataList[i])
         }))
       })
@@ -39,7 +38,7 @@ module.exports = (repo) => {
 
       it('massive read', async function () {
         this.timeout(15000) // add time for ci
-        await Promise.all(_.range(20 * 100).map(async (i) => {
+        await Promise.all(range(20 * 100).map(async (i) => {
           const j = i % dataList.length
           const val = await repo.datastore.get(new Key('hello' + j))
           expect(val).to.be.eql(dataList[j])