+++ test/keystore-test.js
@@ -2,9 +2,7 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
+const { expect } = require('./utils/chai')
 
 module.exports = (repo) => {
   describe('keystore', () => {