+++ test/stat-test.js
@@ -1,10 +1,8 @@
 /* eslint-env mocha */
 'use strict'
 
-const chai = require('chai')
-chai.use(require('dirty-chai'))
-const expect = chai.expect
-const Block = require('ipfs-block')
+const { expect } = require('./utils/chai')
+const Block = require('ipld-block')
 const CID = require('cids')
 
 module.exports = (repo) => {